remove manual user authorziation checks and use 'current_user_dependency' instead

+++ routers/posts.py
@@ -3,7 +3,7 @@ from sqlalchemy import select
 from sqlalchemy.orm import selectinload
 
 import models
-from dependencies import db_dependency
+from dependencies import db_dependency, current_user_dependency
 from schemas import PostCreate, PostResponse, PostUpdate
 
 router = APIRouter()
@@ -25,17 +25,8 @@ async def get_posts(db: db_dependency):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: db_dependency):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
-    new_post = models.Post(title=post.title, content=post.content, user_id=post.user_id)
+async def create_post(post: PostCreate, current_user: current_user_dependency, db: db_dependency):
+    new_post = models.Post(title=post.title, content=post.content, user_id=current_user.id)
     db.add(new_post)
     await db.commit()
     await db.refresh(new_post, attribute_names=["author"])
@@ -61,7 +52,7 @@ async def get_post(post_id: int, db: db_dependency):
 
 
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: db_dependency):
+async def update_post_full(post_id: int, post_data: PostCreate, current_user: current_user_dependency, db: db_dependency):
     result = await db.execute(
         select(models.Post).where(models.Post.id == post_id),
     )
@@ -73,20 +64,15 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: db_dependenc
             detail="Post not found.",
         )
 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post_data.user_id),
+    if post.user_id != current_user.id:
+        raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Not authorized to update this post."
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found.",
-            )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
+    post.user_id = current_user.id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -95,7 +81,7 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: db_dependenc
 
 
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: db_dependency):
+async def update_post_partial(post_id: int, post_data: PostUpdate, current_user: current_user_dependency, db: db_dependency):
     result = await db.execute(
         select(models.Post).where(models.Post.id == post_id),
     )
@@ -105,6 +91,12 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: db_depend
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found.",
+        )
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Not authorized to update this post."
         )
 
     update_data = post_data.model_dump(exclude_unset=True)
@@ -121,7 +113,7 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: db_depend
     "/{post_id}",
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def delete_post(post_id: int, db: db_dependency):
+async def delete_post(post_id: int, current_user: current_user_dependency, db: db_dependency):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
 
@@ -129,6 +121,12 @@ async def delete_post(post_id: int, db: db_dependency):
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found.",
+        )
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail="Not authorized to delete this post."
         )
 
     await db.delete(post)
